product-management/pyproject.toml
[tool.poetry]
name = "product-management"
version = "0.1.0"
description = ""
authors = ["hubaib <hubaibm9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.19"
psycopg = "^3.2.1"
aiokafka = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
